+++ src/Events.sol
@@ -2,7 +2,7 @@
 pragma solidity 0.8.24;
 
 event Claimed(address indexed user, uint128 indexed round, uint128 amount);
-event ClaimedMultiple(address indexed user, uint128[] rounds, uint128 totalAmount);
+event ClaimedMultiple(address indexed user, uint128[] rounds, uint256 totalAmount);
 event SetupRounds(uint256 numOfRounds, uint256 firstClaimTime, uint256 lastClaimTime, uint256 totalAmount);
 event DeadlineUpdated(uint256 indexed newDeadline);
 event Deposited(address indexed operator, uint256 amount);
